CnC_Server/requirements.txt
aiogram==2.18
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
beautifulsoup4==4.10.0
black==21.12b0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cryptography==36.0.1
fastapi==0.72.0
frozenlist==1.3.0
future==0.18.2
greenlet==1.1.2
h11==0.12.0
httptools==0.3.0
idna==3.3
lxml==4.7.1
multidict==6.0.2
mypy-extensions==0.4.3
numpy==1.22.1
pandas==1.3.5
pathspec==0.9.0
platformdirs==2.4.1
psycopg2==2.9.3
pycparser==2.21
pydantic==1.9.0
pyOpenSSL==22.0.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-whois==0.7.3
pytz==2021.3
PyYAML==6.0
requests==2.27.1
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
SQLAlchemy==1.4.29
starlette==0.17.1
tomli==1.2.3
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.17.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
yarl==1.7.2
